ai-powered-new-article-chatbot-main/ai-powered-new-article-chatbot-main/weather-chatbot/pyproject.toml
[project]
name = "weather-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "aman.kadam-coditas",email = "aman.kadam@coditas.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "poethepoet (>=0.37.0,<0.38.0)",
    "bcrypt (==4.3.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "openai (>=2.6.1,<3.0.0)",
    "groq (>=0.33.0,<0.34.0)",
    "pathlib (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "weather_chatbot", from = "src"}]

[tool.poe.tasks]
weatherbot = "poetry run uvicorn src.weather_chatbot.main:app --reload"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
